+++ lizard_progress/views.py
@@ -7,19 +7,24 @@ import shutil
 from matplotlib import figure
 from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
 
+from django.conf import settings
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse
 from django.db import transaction
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.utils import simplejson as json
 from django.views.generic import TemplateView
+from django.views.static import serve
 
 from lizard_map.matplotlib_settings import SCREEN_DPI
 from lizard_map.views import AppView
+from lizard_progress.models import has_access
 from lizard_progress.models import Area
 from lizard_progress.models import Contractor
 from lizard_progress.models import Project
+from lizard_progress.models import MeasurementType
 from lizard_progress.models import ScheduledMeasurement
 from lizard_ui.views import ViewContextMixin
 
@@ -32,6 +37,21 @@ def JsonResponse(ob):
                         mimetype="application/json")
 
 
+def document_root():
+    return getattr(settings, 'LIZARD_PROGRESS_ROOT',
+                   os.path.join(settings.BUILDOUT_DIR, 
+                                'var', 
+                                'lizard_progress'))
+
+
+def make_uploaded_file_path(root, project, contractor, mtype, filename):
+    return os.path.join(root,
+                        project.slug,
+                        contractor.slug,
+                        mtype.slug,
+                        os.path.basename(filename))
+
+
 class View(AppView):
     project_slug = None
     project = None
@@ -41,7 +61,10 @@ class View(AppView):
         self.project_slug = kwargs.get('project_slug', None)
         self.project = get_object_or_404(Project, slug=self.project_slug)
 
-        return super(View, self).dispatch(request, *args, **kwargs)
+        if has_access(request.user, self.project):
+            return super(View, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied()
 
     def upload_url(self):
         return reverse('lizard_progress_uploadview',
@@ -71,17 +94,20 @@ class MapView(View):
         return crumbs
 
     def available_layers(self):
-        return [{
+        logger.debug("Available layers:")
+        layers = [{
                 'name': '%s %s' % (measurement_type.name, contractor.name),
-                'json': '{"contractor_slug":"%s",' +
+                'json': ('{"contractor_slug":"%s",' +
                 '"measurement_type_slug":"%s",' +
-                '"project_slug":"%s"}' %
+                '"project_slug":"%s"}') %
                 (contractor.slug, measurement_type.slug, self.project.slug)
                 }
-                 for contractor in self.project.contractor_set.all()
-                 for measurement_type in
-                 self.project.measurementtype_set.all()]
-
+                  for contractor in self.project.contractor_set.all()
+                  for measurement_type in
+                  self.project.measurementtype_set.all()
+                  if has_access(self.request.user, self.project, contractor)]
+        logger.debug(layers)
+        return layers
 
 class DashboardView(View):
     template_name = 'lizard_progress/dashboard.html'
@@ -100,13 +126,14 @@ class DashboardView(View):
     def areas(self):
         areas = []
         for contractor in Contractor.objects.filter(project=self.project):
-            for area in Area.objects.filter(project=self.project):
-                areas.append((contractor, area,
-                             reverse('lizard_progress_dashboardareaview',
-                                     kwargs={
-                            'contractor_slug': contractor.slug,
-                            'project_slug': self.project.slug,
-                            'area_slug': area.slug})))
+            if has_access(self.request.user, self.project, contractor):
+                for area in Area.objects.filter(project=self.project):
+                    areas.append((contractor, area,
+                                  reverse('lizard_progress_dashboardareaview',
+                                          kwargs={
+                                    'contractor_slug': contractor.slug,
+                                    'project_slug': self.project.slug,
+                                    'area_slug': area.slug})))
 
         return areas
 
@@ -123,7 +150,10 @@ class UploadView(ViewContextMixin, TemplateView):
         self.project_slug = kwargs.get('project_slug', None)
         self.project = get_object_or_404(Project, slug=self.project_slug)
 
-        return super(UploadView, self).dispatch(request, *args, **kwargs)
+        if has_access(request.user, self.project):
+            return super(UploadView, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied()
 
     def upload_url(self):
         return reverse('lizard_progress_uploadview',
@@ -137,29 +167,29 @@ class UploadView(ViewContextMixin, TemplateView):
                 parseresult = parser(path,
                                      project=project,
                                      contractor=contractor)
-
+                
                 if parseresult.success:
                     # Success! Move the file.
                     result_dir = os.path.dirname(parseresult.result_path)
                     if not os.path.exists(result_dir):
                         os.makedirs(result_dir)
                         shutil.move(path, parseresult.result_path)
+                        
+                    return True, {}
+                else:
+                    # Unsuccess. Were there errors? Then set
+                    # them. Note that if there is more than one
+                    # parser, there may still be other parsers
+                    # after this one that are successful, so we
+                    # continue looping.
+                    if parseresult.error:
+                        result = {'error':
+                                      {'details': parseresult.error}}
 
-                        return True, {}
-                    else:
-                        # Unsuccess. Were there errors? Then set
-                        # them. Note that if there is more than one
-                        # parser, there may still be other parsers
-                        # after this one that are successful, so we
-                        # continue looping.
-                        if parseresult.error:
-                            result = {'error':
-                                          {'details': parseresult.error}}
-
-                        # We raise a dummy exception so that
-                        # commit_on_success doesn't commit whatever
-                        # was done to our database in the meantime.
-                        raise DummyException()
+                    # We raise a dummy exception so that
+                    # commit_on_success doesn't commit whatever
+                    # was done to our database in the meantime.
+                    raise DummyException()
         except DummyException:
             pass
 
@@ -222,6 +252,9 @@ class DashboardAreaView(View):
         self.contractor = get_object_or_404(Contractor,
                                             project=self.project,
                                             slug=self.contractor_slug)
+        
+        if not has_access(request.user, self.project, self.contractor):
+            raise PermissionDenied()
 
         return (super(DashboardAreaView, self).
                 dispatch(request, *args, **kwargs))
@@ -261,8 +294,10 @@ def dashboard_graph(request, project_slug, contractor_slug,
     contractor = get_object_or_404(Contractor, project=project,
                                    slug=contractor_slug)
 
-    # XXX
-    fig = ScreenFigure(600, 300)  # in pixels
+    if not has_access(request.user, project, contractor):
+        raise PermissionDenied()
+
+    fig = ScreenFigure(600, 350)  # in pixels
     fig.text(0.5, 0.85, 'Uitgevoerde werkzaamheden', fontsize=14, ha='center')
     fig.subplots_adjust(left=0.05, right=0.95)  # smaller margins
     y_title = -0.2  # a bit lower
@@ -315,4 +350,60 @@ def dashboard_graph(request, project_slug, contractor_slug,
     response = HttpResponse(content_type='image/png')
     canvas = FigureCanvas(fig)
     canvas.print_png(response)
+    return response
+
+
+@login_required
+def protected_file_download(request, project_slug, contractor_slug,
+                                   measurement_type_slug, filename):
+    """
+    We need our own file_download view because contractors can only see their
+    own files, and the URLs of other contractor's files are easy to guess.
+
+    Copied and adapted from deltaportaal, which has a more generic
+    example if you're looking for one. It is for Apache.
+
+    The one below works for both Apache (untested) and Nginx.  I used
+    the docs at http://wiki.nginx.org/XSendfile for the Nginx
+    configuration.  Basically, Nginx serves /protected/ from the
+    document root at BUILDOUT_DIR+'var', and we x-accel-redirect
+    there. Also see the bit of nginx conf in hdsr's etc/nginx.conf.in.
+    """
+
+    project = get_object_or_404(Project, slug=project_slug)
+    contractor = get_object_or_404(Contractor, slug=contractor_slug,
+                                   project=project)
+    mtype = get_object_or_404(MeasurementType, slug=measurement_type_slug,
+                              project=project)
+
+    logger.debug("Incoming programfile request for %s", filename)
+
+    if '/' in filename or '\\' in filename:
+        # Trickery
+        logger.warn("Returned 403 on suspect path %s" % (filename,))
+
+    file_path = make_uploaded_file_path(document_root(), project, contractor,
+                                        mtype, filename)
+    nginx_path = make_uploaded_file_path('/protected', project, contractor,
+                                        mtype, filename)
+
+    if not has_access(request.user, project, contractor):
+        logger.warn("Not allowed to access %s", filename)
+        return HttpResponseForbidden()
+
+    # This is where the magic takes place.
+    response = HttpResponse()
+    response['X-Sendfile'] = file_path # Apache
+    response['X-Accel-Redirect'] = nginx_path # Nginx
+
+    # Unset the Content-Type as to allow for the webserver
+    # to determine it.
+    response['Content-Type'] = ''
+    # TODO: ... or USE_IIS:...
+    if settings.DEBUG or not platform.system() == 'Linux':
+        logger.debug(
+            "With DEBUG off, we'd serve the programfile via apache: \n%s",
+            response)
+        return serve(request, file_path, '/')
+    logger.debug("Serving programfile %s via apache.", fullpath)
     return response