+++ CHANGES.rst
@@ -6,3 +6,9 @@ Changelog of lizard-progress
 ----------------
 
 - Initial library skeleton created by nensskel.  [Remco Gerlich]
+
+- A lot of stuff works; we can have multiple projects, multiple contractors,
+  subareas, measurement types, we can schedule measurements and upload files
+  that can be parsed. We can show layers and dashboard graphs and serve back the
+  files, only to the right contractors or superusers. I'm marking this at 0.1 for
+  no particular reason.